+++ cvat-ui/src/components/annotation-page/top-bar/player-navigation.tsx
@@ -24,6 +24,7 @@ interface Props {
     frameNumber: number;
     frameFilename: string;
     frameDeleted: boolean;
+    deleteFrameAvailable: boolean;
     deleteFrameShortcut: string;
     focusFrameInputShortcut: string;
     inputFrameRef: React.RefObject<Input>;
@@ -52,6 +53,7 @@ function PlayerNavigation(props: Props): JSX.Element {
         onDeleteFrame,
         onRestoreFrame,
         switchNavigationBlocked,
+        deleteFrameAvailable,
     } = props;
 
     const [frameInputValue, setFrameInputValue] = useState<number>(frameNumber);
@@ -81,6 +83,15 @@ function PlayerNavigation(props: Props): JSX.Element {
             });
         }
     }, [playing, frameNumber]);
+    const deleteFrameIcon = !frameDeleted ? (
+        <CVATTooltip title={`Delete the frame ${deleteFrameShortcut}`}>
+            <DeleteOutlined className='cvat-player-delete-frame' onClick={showDeleteFrameDialog} />
+        </CVATTooltip>
+    ) : (
+        <CVATTooltip title='Restore the frame'>
+            <Icon className='cvat-player-restore-frame' onClick={onRestoreFrame} component={RestoreIcon} />
+        </CVATTooltip>
+    );
 
     return (
         <>
@@ -106,15 +117,9 @@ function PlayerNavigation(props: Props): JSX.Element {
                         <CVATTooltip title='Create frame URL'>
                             <LinkOutlined className='cvat-player-frame-url-icon' onClick={onURLIconClick} />
                         </CVATTooltip>
-                        { (!frameDeleted) ? (
-                            <CVATTooltip title={`Delete the frame ${deleteFrameShortcut}`}>
-                                <DeleteOutlined className='cvat-player-delete-frame' onClick={showDeleteFrameDialog} />
-                            </CVATTooltip>
-                        ) : (
-                            <CVATTooltip title='Restore the frame'>
-                                <Icon className='cvat-player-restore-frame' onClick={onRestoreFrame} component={RestoreIcon} />
-                            </CVATTooltip>
-                        )}
+                        {
+                            deleteFrameAvailable && deleteFrameIcon
+                        }
                     </Col>
                 </Row>
             </Col>