+++ cvat-ui/src/components/annotation-page/top-bar/top-bar.tsx
@@ -44,6 +44,7 @@ interface Props {
     focusFrameInputShortcut: string;
     activeControl: ActiveControl;
     toolsBlockerState: ToolsBlockerState;
+    deleteFrameAvailable: boolean;
     changeWorkspace(workspace: Workspace): void;
     showStatistics(): void;
     showFilters(): void;
@@ -120,6 +121,7 @@ export default function AnnotationTopBarComponent(props: Props): JSX.Element {
         onFinishDraw,
         onSwitchToolsBlockerState,
         onDeleteFrame,
+        deleteFrameAvailable,
         onRestoreFrame,
         switchNavigationBlocked,
         jobInstance,
@@ -178,6 +180,7 @@ export default function AnnotationTopBarComponent(props: Props): JSX.Element {
             onDeleteFrame={onDeleteFrame}
             onRestoreFrame={onRestoreFrame}
             switchNavigationBlocked={switchNavigationBlocked}
+            deleteFrameAvailable={deleteFrameAvailable}
         />
     ), 10]);