+++ cvat-ui/src/reducers/index.ts
@@ -6,7 +6,7 @@
 import { Canvas3d } from 'cvat-canvas3d/src/typescript/canvas3d';
 import { Canvas, RectDrawingMethod, CuboidDrawingMethod } from 'cvat-canvas-wrapper';
 import {
-    Webhook, MLModel, ModelProvider, Organization,
+    Webhook, MLModel, ModelProvider, Organization, QualityReport, QualityConflict, QualitySettings, FramesMetaData,
 } from 'cvat-core-wrapper';
 import { IntelligentScissors } from 'utils/opencv-wrapper/intelligent-scissors';
 import { KeyMap } from 'utils/mousetrap-react';
@@ -92,6 +92,11 @@ export interface JobsState {
     current: Job[];
     previews: {
         [index: number]: Preview;
+    };
+    activities: {
+        deletes: {
+            [tid: number]: boolean;
+        };
     };
 }
 
@@ -454,6 +459,8 @@ export interface NotificationsState {
         jobs: {
             updating: null | ErrorState;
             fetching: null | ErrorState;
+            creating: null | ErrorState;
+            deleting: null | ErrorState;
         };
         formats: {
             fetching: null | ErrorState;
@@ -545,6 +552,11 @@ export interface NotificationsState {
             updating: null | ErrorState;
             deleting: null | ErrorState;
         };
+        analytics: {
+            fetching: null | ErrorState;
+            fetchingSettings: null | ErrorState;
+            updatingSettings: null | ErrorState;
+        }
     };
     messages: {
         tasks: {
@@ -663,6 +675,8 @@ export interface AnnotationState {
         openTime: null | number;
         labels: any[];
         requestedId: number | null;
+        groundTruthJobId: number | null;
+        groundTruthJobFramesMeta: FramesMetaData | null;
         instance: any | null | undefined;
         attributes: Record<number, any[]>;
         fetching: boolean;
@@ -696,8 +710,10 @@ export interface AnnotationState {
         activatedStateID: number | null;
         activatedElementID: number | null;
         activatedAttributeID: number | null;
+        highlightedConflict: QualityConflict | null;
         collapsed: Record<number, boolean>;
         collapsedAll: boolean;
+        statesSources: number[];
         states: any[];
         filters: any[];
         resetGroupFlag: boolean;
@@ -807,6 +823,7 @@ export interface ShapesSettingsState {
     outlineColor: string;
     showBitmap: boolean;
     showProjections: boolean;
+    showGroundTruth: boolean;
 }
 
 export interface SettingsState {
@@ -840,6 +857,8 @@ export interface ReviewState {
     newIssuePosition: number[] | null;
     issuesHidden: boolean;
     issuesResolvedHidden: boolean;
+    conflicts: QualityConflict[];
+    frameConflicts: QualityConflict[];
     fetching: {
         jobId: number | null;
         issueId: number | null;
@@ -875,6 +894,26 @@ export interface WebhooksState {
     query: WebhooksQuery;
 }
 
+export interface QualityQuery {
+    taskId: number | null;
+    jobId: number | null;
+    parentId: number | null;
+}
+
+export interface AnalyticsState {
+    fetching: boolean;
+    quality: {
+        tasksReports: QualityReport[];
+        jobsReports: QualityReport[];
+        query: QualityQuery;
+        settings: {
+            modalVisible: boolean;
+            current: QualitySettings | null;
+            fetching: boolean;
+        }
+    }
+}
+
 export interface CombinedState {
     auth: AuthState;
     projects: ProjectsState;
@@ -895,6 +934,7 @@ export interface CombinedState {
     cloudStorages: CloudStoragesState;
     organizations: OrganizationState;
     webhooks: WebhooksState;
+    analytics: AnalyticsState;
 }
 
 export interface Indexable {