+++ cvat/apps/dataset_manager/task.py
@@ -24,7 +24,6 @@ from cvat.apps.dataset_manager.bindings import TaskData, JobData, CvatImportErro
 from cvat.apps.dataset_manager.formats.registry import make_exporter, make_importer
 from cvat.apps.dataset_manager.util import add_prefetch_fields, bulk_create, get_cached
 
-
 class dotdict(OrderedDict):
     """dot.notation access to dictionary attributes"""
     __getattr__ = OrderedDict.get